backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1
langchain==1.2.7
langgraph==1.0.7
langchain-openai==1.1.7
chromadb==0.4.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.1.0
email-validator==2.2.0
python-dateutil==2.8.2
pypdf==4.0.1
python-docx==1.1.0
pillow==10.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
httpx==0.26.0
loguru==0.7.2
aiofiles==24.1.0
numpy==1.26.4
